+++ src/svmlite/__init__.py
@@ -1,6 +1,6 @@
 """SVMLite: A lightweight SVM implementation from scratch."""
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 from .svm import SVCLite
 from .utils import StandardScalerLite